Make NAmespaceAndName standalone class
